[ADD - Client]: update product added
