+++ client/src/dashboard/products/helpers/index.js
@@ -1,1 +1,1 @@
-export * from './parseDataCreate';
+export * from './parseData';
